pyproject.toml
[tool.poetry]
name = "books_shop"
version = "0.1.0"
description = "digit unicorn lessons"
authors = ["bogdankozlovsky <bogdan.k@datarails.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.9"
djangorestframework = "3.13.1"
djoser = "2.1.0"
channels = "3.0.4"
django-debug-toolbar = "^3.2.4"
django-filter = "^21.1"
django-guardian = "^2.4.0"
django-extensions = "^3.1.5"
Pillow = "^9.0.0"
jupyter = "^1.0.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
